% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddActivityAssay.R
\name{AddActivityAssay}
\alias{AddActivityAssay}
\title{This function adds a named assay 'activity' to a Seurat object.}
\usage{
AddActivityAssay(seu, vp_mat)
}
\description{
This function adds a named assay 'activity' to a Seurat object.
}
