% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddMarkerAssay.R
\name{AddMarkerAssay}
\alias{AddMarkerAssay}
\title{This function adds a named assay 'activity' to a Seurat object.}
\usage{
AddMarkerAssay(seu, marker.mat)
}
\description{
This function adds a named assay 'activity' to a Seurat object.
}
