% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aracne2Regulon.R
\name{Aracne2Regulon2}
\alias{Aracne2Regulon2}
\title{This function generate the regulon}
\usage{
Aracne2Regulon2(
  network_dir_path,
  cutoff = 50,
  adaptive = FALSE,
  eliminate = TRUE,
  wm = NULL
)
}
\description{
This function generate the regulon
}
\author{
Junqiang Wang
}
