% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aracne2Regulon.R
\name{Aracne2RegulonMono}
\alias{Aracne2RegulonMono}
\title{This function generate the regulon}
\usage{
Aracne2RegulonMono(
  log2mat.file,
  network.file,
  cutoff = 50,
  adaptive = FALSE,
  eliminate = TRUE,
  wm = NULL
)
}
\description{
This function generate the regulon
}
\author{
Junqiang Wang
}
