% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aREA_GSEA_PRESTO_CancerHallMarks.R
\name{aREA_GSEA}
\alias{aREA_GSEA}
\title{this function do gsea enrichment analysis}
\usage{
aREA_GSEA(current.geneset, current.sig, uniform.wts = TRUE, tcm.size = NULL)
}
\description{
this function do gsea enrichment analysis
}
\author{
Junqing Wang
}
