% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aREA_GSEA_PRESTO_CancerHallMarks.R
\name{aREA_PRESTO_CancerHallMarkers}
\alias{aREA_PRESTO_CancerHallMarkers}
\title{this function do gsea enrichment analysis}
\usage{
aREA_PRESTO_CancerHallMarkers(
  seu,
  cluster.by = "seurat_clusters",
  species = c("human", "mouse"),
  signature.cluster = NULL,
  groups.use = NULL,
  reverse.clusters = c(TRUE, FALSE),
  assay = "RNA",
  slot = c("scale.data", "counts", "data")
)
}
\description{
this function do gsea enrichment analysis
}
\author{
Junqing Wang
}
