% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aecdf.r
\name{cdfInteg}
\alias{cdfInteg}
\title{Integration based on CDF (AOC)}
\usage{
cdfInteg(x, xlim = NULL)
}
\arguments{
\item{x}{Numeric vector, matrix or list of vectors or matrixes}

\item{xlim}{Numeric vector of 2 elements indicating the range where to perform the integration}
}
\description{
This function integrates a distribution of scores based on the area over the CDF curve
}
\details{
This function computes the area over the curve for the vector o columns of the matrix provided as input
}
