% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoComplexHeatmap.R
\name{DoComplexHeatmap}
\alias{DoComplexHeatmap}
\title{Accustomed Heatmap Plot of the Top Differentially Activate Proteins}
\usage{
DoComplexHeatmap(
  seu,
  cluster.by = "seurat_clusters",
  column_split_option = c("YES", "NO"),
  assay = "Activity",
  slot = c("data", "scale.data", "data"),
  reverse.clusters = c(FALSE, TRUE),
  ann.top.df.columns = c("seurat_clusters", "Phase"),
  data.type = c("expression", "activity"),
  color_quantile = 0.01,
  fontsize.column.title.gp = 7,
  fontsize.row.names.gp = 7
)
}
\arguments{
\item{seu}{the seurat markers}

\item{cluster.by}{the column names indicating the column that is used in the analysis}
}
\description{
Accustomed Heatmap Plot of the Top Differentially Activate Proteins
}
