% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneNameConverterMouse.R
\name{Ens2SymblMat_Mouse}
\alias{Ens2SymblMat_Mouse}
\title{this function convert the mouse gene names}
\usage{
Ens2SymblMat_Mouse(mat)
}
\arguments{
\item{mat}{The matrix with the rownames of}
}
\description{
this function convert the mouse gene names
}
\author{
Junqing Wang
}
