% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entrez2HGNC_Biomart.R
\name{Ensembl2SymbolMat_Human}
\alias{Ensembl2SymbolMat_Human}
\title{This function convert the human Entrez names to symbl}
\usage{
Ensembl2SymbolMat_Human(mat)
}
\arguments{
\item{mat}{matrix, for example viper matrix}
}
\value{
A matrix associated with gene symbol
}
\description{
This function convert the human Entrez names to symbl
}
\author{
Junqiang Wang
}
