% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoufferIntegrateSeuClusterRow.R
\name{GetActivityMarkersStouffer}
\alias{GetActivityMarkersStouffer}
\alias{StoufferIntegrateSeuClusterMarkers}
\title{This function do Stouffer integration and identify the markers}
\usage{
GetActivityMarkersStouffer(
  seu,
  top.n = 25,
  assay = "Activity",
  slot = "data",
  output.format = c("seurat", "matrix"),
  reverse.clusters = FALSE,
  order.cluster = NULL,
  cluster.by = NULL
)
}
\arguments{
\item{seu}{The seurat object}
}
\value{
Integreated NES
}
\description{
This function do Stouffer integration and identify the markers
}
