% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExpressionMarkers.R
\name{GetExpressionMarkers}
\alias{GetExpressionMarkers}
\title{This function generate an accustomed Heatmap Plot of the Top Differentially Activate Proteins}
\usage{
GetExpressionMarkers(
  seu,
  cluster.by = "seurat_clusters",
  reverse.clusters = c(FALSE, TRUE),
  ident.1 = NULL,
  ident.2 = NULL,
  top.n = 25,
  assay = "RNA",
  slot.use = c("scale.data", "data", "counts"),
  only.pos = TRUE,
  test.use = "wilcox",
  min.pct.threshold = 0.25,
  myAUC.cutoff = 0.5,
  avg_logfc.threshold = 0,
  avg_diff.threshold = 0,
  p_val_adj.threshold = 0.01
)
}
\arguments{
\item{seu}{seurat object}

\item{cluster.by}{the ID in seu@meata.data}

\item{reverse.clusters}{whether the order of the cluster id shoud be reversed}

\item{top.n}{return the top.n dif features}

\item{assay}{The assay that is used}

\item{cluster.order}{the order of the cluster}

\item{groups.use}{the groups that are used run the presto}

\item{slot}{the slot of the assay that is used}
}
\value{
return the markers and store the markers in a new assay
}
\description{
This function generate an accustomed Heatmap Plot of the Top Differentially Activate Proteins
}
\author{
Junqiang Wang
}
