% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMarkers.R
\name{GetMarkers}
\alias{GetMarkers}
\title{This function generate an accustomed Heatmap Plot of the Top Differentially Activate Proteins}
\usage{
GetMarkers(
  seu,
  cluster.by = "seurat_clusters",
  reverse.clusters = c(FALSE, TRUE),
  ident.1 = NULL,
  ident.2 = NULL,
  top.n = 25,
  assay = "Activity",
  slot = c("data", "scale.data", "counts"),
  only.pos = TRUE,
  min.pct.threshold = 0.5,
  logfc.threshold = 0.25,
  p_val_adj.threshold = 0.01,
  test.use = c("roc", "wilcox")
)
}
\arguments{
\item{seu}{seurat object}

\item{cluster.by}{the ID in seu@meata.data}

\item{reverse.clusters}{whether the order of the cluster id shoud be reversed}

\item{top.n}{return the top.n dif features}

\item{assay}{The assay that is used}

\item{slot}{the slot of the assay that is used}

\item{cluster.order}{the order of the cluster}

\item{groups.use}{the groups that are used run the presto}
}
\value{
return the markers and store the markers in a new assay
}
\description{
This function generate an accustomed Heatmap Plot of the Top Differentially Activate Proteins
}
\author{
Junqiang Wang
}
