% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMarkersStouffer.R
\name{GetMarkersStouffer}
\alias{GetMarkersStouffer}
\title{This function generates the markers of the Top Differentially Activate Proteins by using the Stouffer integration method}
\usage{
GetMarkersStouffer(
  seu,
  cluster.by = "seurat_clusters",
  reverse.clusters = c(FALSE, TRUE),
  groups.use = NULL,
  top.n = 25,
  marker.type = c("both", "top", "bottom"),
  assay = "Activity",
  slot = c("data", "scale.data", "counts")
)
}
\arguments{
\item{seu}{seurat object}

\item{cluster.by}{the ID in seu@meata.data}

\item{reverse.clusters}{whether the order of the cluster id shoud be reversed}

\item{top.n}{return the top.n dif features}

\item{assay}{The assay that is used}

\item{slot}{the slot of the assay that is used}

\item{cluster.order}{the order of the cluster}
}
\value{
return the markers and store the markers in a new assay
}
\description{
This function generates the markers of the Top Differentially Activate Proteins by using the Stouffer integration method
}
\author{
Junqiang Wang
}
