% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.r
\name{gsea}
\alias{gsea}
\title{Gene Set Enrichment Analysis}
\usage{
gsea(
  signature,
  geneset,
  score = 1,
  twoTails = FALSE,
  pout = TRUE,
  per = 0,
  alternative = c("two.sided", "greater", "less"),
  colSig = c(0.45, 0.15, 0.3, 1),
  colHit = c(0.58, 0.05, 0.1, 2),
  ylim = NULL,
  axes = TRUE,
  xlab = "Signature",
  ylab = "ES",
  lwd = 1,
  maxhit = 0,
  ...
)
}
\arguments{
\item{signature}{Numeric vector containing the gene expression signature}

\item{geneset}{Vector of character strings indicating the genes in the geneset, or named numeric vector for 2-tail analysis}

\item{score}{Number indicating the exponent score for GSEA}

\item{twoTails}{Logical, whether the ES should be computed as the different between each tail ES}

\item{pout}{Logical, whether a plot should be generated}

\item{per}{Integer indicating the number of permutations for p-value estimation}

\item{alternative}{Character string indicating the tail to test for statistical significance}

\item{colSig}{Vector indicating the colors for the signature, including hue for negative values, hue for positive values, mate value and gama}

\item{colHit}{Vector indicating the colors for the hits, including hue for negative values, hue for positive values, mate value and gama}

\item{ylim}{Optional numeric vector indicating the limits for the y-axis}

\item{axes}{Logical, whether axis should be ploted}

\item{xlab}{Character string indicating the label for the x-axis}

\item{ylab}{Character string indicating the label for the y-axis}

\item{lwd}{Number indicating the line width for the enrichment plot}

\item{maxhit}{Integer indicating the maximum number of hits to plot, 0 means all}

\item{...}{Additional parameters to pass to plot function}
}
\value{
List of results and plot
}
\description{
This function performs gene eset enrichment analysis and plots
}
