% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Count2logCPM.R
\name{Log2CPM}
\alias{Log2CPM}
\title{This function convert count matrix to log2CPM matrix}
\usage{
Log2CPM(count)
}
\arguments{
\item{count}{The count matrix}
}
\value{
The log2CPM matrix
}
\description{
This function convert count matrix to log2CPM matrix
}
\author{
Junqiang Wang
}
