% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mat2AracneMat.R
\name{Mat2Aracne}
\alias{Mat2Aracne}
\title{This function generate the arance mat files from matrix}
\usage{
Mat2Aracne(
  mat,
  outfile.name.suffix,
  log2CPM = c("On", "Off"),
  write.output = c(TRUE, FALSE),
  sample.use = "On",
  r.cutoff = 0.1,
  sample.size = 500
)
}
\arguments{
\item{log2CPM}{if on, use log2CPM transformation}

\item{write.output}{if TRUE, output will write to files}
}
\description{
This function generate the arance mat files from matrix
}
