% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entrez2HGNC_Biomart.R
\name{NameConvertor}
\alias{NameConvertor}
\title{This function generate a one-to-one name list including entrez and hgnc_symbol}
\usage{
NameConvertor()
}
\description{
This function generate a one-to-one name list including entrez and hgnc_symbol
}
\author{
Junqiang Wang
}
