% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niceExponent.r
\name{niceExponent}
\alias{niceExponent}
\title{Nice Exponential representations of scientific notation}
\usage{
niceExponent(
  x,
  drop.1 = TRUE,
  sub10 = "10",
  digits = 2,
  digits.fuzz,
  lab.type = c("plotmath", "latex"),
  lab.sep = c("cdot", "times")
)
}
\arguments{
\item{x}{Numeric vector}

\item{drop.1}{Logical, whether 1 in 1 x type of representatons should be dropped}

\item{sub10}{Either logical, "10", a non-negative integer or a length 2 integer vector, indicating if some expression should be formatted traditionally, when integer, all expression before the integer are simplified. when a 2 elements vector, all between the indicated range are simplified}

\item{digits}{Number of significant digits}

\item{lab.type}{Character string indicating how the result should look like, either plotmath or latex}

\item{lab.sep}{Character separator between mantissa and exponent}
}
\value{
Vector of formated numbers
}
\description{
This function generates a plotmath or latex representation of scientific notation
}
