% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot2viperSimilarity.R
\name{Plot2viperSimilarity}
\alias{Plot2viperSimilarity}
\title{This function compares the viper similarity ofphenotype datasets}
\usage{
Plot2viperSimilarity(
  pa.dat.query,
  pa.dat.ref,
  nn = 50,
  method = c("two.sided", "greater", "less")
)
}
\arguments{
\item{pa.dat.query}{The query dataset}

\item{pa.dat.ref}{The reference dataset}

\item{nn}{The param in function viperSimilarity}

\item{method}{The param in function viperSimilarity}
}
\description{
This function compares the viper similarity ofphenotype datasets
}
\author{
Junqiang Wang
}
