% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleGSEA_PRESTO_CancerHallMarks.R,
%   R/aREA_GSEA_PRESTO_CancerHallMarks.R
\name{PlotHeatmapGSEA}
\alias{PlotHeatmapGSEA}
\title{This function do heatmap plot of the gsea analysis result}
\usage{
PlotHeatmapGSEA(mat, col_fun_range = c(-10, 0, 10))

PlotHeatmapGSEA(mat, col_fun_range = c(-10, 0, 10))
}
\description{
This function do heatmap plot of the gsea analysis result

This function do heatmap plot of the gsea analysis result
}
\author{
Junqiang

Junqiang
}
