% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoComplexHeatmapMarkers.R
\name{PlotSeuMarkers}
\alias{PlotSeuMarkers}
\title{Accustomed Heatmap Plot of the Top Differentially Activate Proteins}
\usage{
PlotSeuMarkers(
  seu,
  cluster.by = "seurat_clusters",
  column_split_option = c("YES", "NO"),
  assay = "Marker",
  slot = c("data", "scale.data", "count"),
  reverse.clusters = c(FALSE, TRUE),
  ann.top.df.columns = c("seurat_clusters", "Phase"),
  legend.name = c("Expression", "Activity"),
  color_quantile = 0.01,
  fontsize.column.title.gp = 7,
  fontsize.row.names.gp = 7,
  column_names_rot = 0,
  column_title_rot = 0,
  order.cluster = NULL,
  rowAnnotation = NULL
)
}
\arguments{
\item{seu}{the seurat markers}

\item{cluster.by}{the column names indicating the column that is used in the analysis}
}
\description{
Accustomed Heatmap Plot of the Top Differentially Activate Proteins
}
