% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrestoMarkers.R
\name{PrestoMarkers}
\alias{PrestoMarkers}
\title{This function generate an accustomed Heatmap Plot of the Top Differentially Activate Proteins}
\usage{
PrestoMarkers(
  seu,
  cluster.by = "seurat_clusters",
  reverse.clusters = c(FALSE, TRUE),
  groups.use = NULL,
  top.n = 25,
  assay = "Activity",
  slot = c("data", "scale.data", "counts"),
  pval.cutoff = 1,
  padj.cutoff = 0.01,
  logFC.cutoff = 0.25
)
}
\arguments{
\item{seu}{seurat object}

\item{cluster.by}{the ID in seu@meata.data}

\item{reverse.clusters}{whether the order of the cluster id shoud be reversed}

\item{groups.use}{the groups that are used run the presto}

\item{top.n}{return the top.n dif features}

\item{assay}{The assay that is used}

\item{slot}{the slot of the assay that is used}

\item{logFC.cutoff}{the cutoff of the presto}

\item{cluster.order}{the order of the cluster}
}
\value{
return the markers and store the markers in a new assay
}
\description{
This function generate an accustomed Heatmap Plot of the Top Differentially Activate Proteins
}
\author{
Junqiang Wang
}
