% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoBulkNcells.R
\name{PseudoBulkNcells}
\alias{PseudoBulkNcells}
\title{This function create the pseudobulk using N cells}
\usage{
PseudoBulkNcells(
  mat,
  N.cells = 10,
  N.pseudoBulk = 10,
  suffix.pseudoBulk = NULL,
  replace = TRUE
)
}
\arguments{
\item{mat}{matrix of single cells}

\item{N}{number of cells}
}
\description{
This function create the pseudobulk using N cells
}
