% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegulonGeneration.R
\name{RegulGeneration}
\alias{RegulGeneration}
\title{This function generate the regulon from the output of the ARACNe}
\usage{
RegulGeneration(log2mat_dir, network_dir, cutoff = 50)
}
\arguments{
\item{network_dir}{the directory of the ARACNe network}

\item{cutoff}{the param in function pruneRegulon}

\item{lgo2mat_dir}{the directory of the expression data}
}
\value{
pruned regulon
}
\description{
This function generate the regulon from the output of the ARACNe
}
\author{
Junqiang Wang
}
