% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuClusteringVPsimilarity.R
\name{SeuClusteringVPsimilarity}
\alias{SeuClusteringVPsimilarity}
\title{This function do clustering analysis using viper similarity as distance}
\usage{
SeuClusteringVPsimilarity(
  seu,
  nn = 50,
  method = c("two.sided", "greater", "less"),
  my.dims = 1:50,
  features = rownames(seu),
  k.param = 20,
  prune.SNN = 1/15,
  resolution = 0.3,
  graph.name = "snn"
)
}
\arguments{
\item{seu}{The seu object of protein activity}
}
\value{
The seurat object associated with cluster info
}
\description{
This function do clustering analysis using viper similarity as distance
}
\author{
Junqiang Wang
}
