% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuMerge.R
\name{SeuMerge}
\alias{SeuMerge}
\title{This function merge the seurat object}
\usage{
SeuMerge(seu.list, add.cell.ids = NULL, project = "Seu")
}
\arguments{
\item{seu.list}{list of seurat object}
}
\description{
This function merge the seurat object
}
