% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuPAClustering.R
\name{SeuPAClustering}
\alias{SeuPAClustering}
\title{This function do Clustering Analysis. Different distance methods are included, Euclidean, Viper Similairty and Correlation distance. Please choose the assay abd the slot to do the analysis.}
\usage{
SeuPAClustering(
  seu,
  ndims = 30,
  dims = 10,
  resolution = 0.8,
  assay = c("Activity", "RNA"),
  slot = c("data", "counts", "scaled.data"),
  feature.sd.cutoff = 0,
  distance = c("default", "viper_similarity", "cor"),
  k.param = 20,
  prune.SNN = 1/15,
  nn = 50
)
}
\arguments{
\item{seu}{The seu expression object
res=0.6 for PA}
}
\value{
The seu results
}
\description{
This function do Clustering Analysis. Different distance methods are included, Euclidean, Viper Similairty and Correlation distance. Please choose the assay abd the slot to do the analysis.
}
\author{
Junqiang Wang
}
