% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pseudobulk.R
\name{seuPseudoBulk}
\alias{seuPseudoBulk}
\title{PseudobulkCountGen}
\usage{
seuPseudoBulk(seu, cluster.by)
}
\arguments{
\item{seu}{The seurat object}

\item{cluster.by}{the colnames of the seu@meta.data}
}
\value{
A count matrix
}
\description{
generate the Pseudobulk counts from scRNA-Seq counts
}
