% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuSCTpipeline.R
\name{SeuSCTpipeline}
\alias{SeuSCTpipeline}
\title{This function run the seurat do SCTransform}
\usage{
SeuSCTpipeline(
  seu,
  species = c("Human", "Mouse"),
  method = c("default", "glmGamPoi"),
  vars.to.regress = NULL,
  ndims = 50,
  dims = 50,
  resolution = 0.8
)
}
\arguments{
\item{seu}{The seu expression object}

\item{method}{The method used for SCTransform}

\item{vars.to.regress}{The variables used in SCTransform, for example, "percent.mt", "nCount_RNA"}
}
\value{
The seu results
}
\description{
This function run the seurat do SCTransform
}
\author{
Junqiang Wang
}
