% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuStandardPipeline.R
\name{SeuStandardPipeline}
\alias{SeuStandardPipeline}
\title{This function run the seurat do SCTransform}
\usage{
SeuStandardPipeline(
  seu,
  species = c("Human", "Mouse"),
  ndims = 50,
  dims = 50,
  resolution = 0.8
)
}
\arguments{
\item{seu}{The seu expression object}
}
\value{
The seu results
}
\description{
This function run the seurat do SCTransform
}
\author{
Junqiang Wang
}
