% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuViper.R
\name{SeuViper}
\alias{SeuViper}
\title{This function generate the protein activity and store the protein activity in the}
\usage{
SeuViper(
  seu,
  regulon,
  ref = NULL,
  regression = FALSE,
  Rank.transformation = FALSE,
  sig.method = c("viperSignature", "Z-transform", "RZ-transform"),
  drop.na = c(TRUE, FALSE),
  mvws = 1
)
}
\arguments{
\item{seu}{The seu expression object}
}
\value{
The seu results
}
\description{
This function generate the protein activity and store the protein activity in the
}
\author{
Junqiang Wang
}
