% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuViper.R
\name{SeuViperInternal}
\alias{SeuViperInternal}
\title{This function generate the protein activity and store the protein activity in the}
\usage{
SeuViperInternal(seu, regulon)
}
\arguments{
\item{seu}{The seu expression object}
}
\value{
The seu results
}
\description{
This function generate the protein activity and store the protein activity in the
}
\author{
Junqiang Wang
}
