% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureGenExternal.R
\name{SignaureGenExternal}
\alias{SignaureGenExternal}
\title{This function generates the signatue of an expression mat using the external expression mat}
\usage{
SignaureGenExternal(exp, exp.ref, useLogTransform = TRUE)
}
\arguments{
\item{exp}{The matrix}

\item{exp.ref}{The external matrix as centroid reference}

\item{useLogTransform}{Whether process log transformation on the input datasets}
}
\description{
This function generates the signatue of an expression mat using the external expression mat
}
\author{
Junqiang Wang
}
