% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleGSEA_PRESTO_CancerHallMarks.R
\name{SimpleGSEA_PRESTO_CancerHallMarkers_Ref}
\alias{SimpleGSEA_PRESTO_CancerHallMarkers_Ref}
\title{this function do gsea enrichment analysis using fixed reference}
\usage{
SimpleGSEA_PRESTO_CancerHallMarkers_Ref(
  seu,
  cluster.by = "seurat_clusters",
  species = c("human", "mouse"),
  signature.cluster = NULL,
  groups.use = NULL,
  ref.cluster = NULL,
  reverse.clusters = c(TRUE, FALSE),
  assay = "RNA",
  slot = c("scale.data", "counts", "data")
)
}
\arguments{
\item{ref.cluster}{The reference cluster}
}
\description{
this function do gsea enrichment analysis using fixed reference
}
\author{
Junqing Wang
}
