% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoufferIntegrateSeuClusterRow.R
\name{StoufferIntegrateSeu1ClusterRows}
\alias{StoufferIntegrateSeu1ClusterRows}
\title{This function do Stouffer integration of the PA in 1 cluster by rows}
\usage{
StoufferIntegrateSeu1ClusterRows(
  seu,
  nn = 25,
  assay = "Activity",
  slot = "data",
  cluster.by = "PA.seurat_clusters",
  sub_clusters = "Null"
)
}
\arguments{
\item{seu}{The seurat object}
}
\value{
Integreated NES
}
\description{
This function do Stouffer integration of the PA in 1 cluster by rows
}
