% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Entrez2HGNC_Biomart.R
\name{Symbol2EntrezMat}
\alias{Symbol2EntrezMat}
\title{This function convert the human symbl to Entrez names}
\usage{
Symbol2EntrezMat(mat)
}
\arguments{
\item{mat}{matrix, for example viper matrix}
}
\value{
A matrix associated with gene symbol
}
\description{
This function convert the human symbl to Entrez names
}
\author{
Junqiang Wang
}
